requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 
pytest==8.4.0
email-validator
httpx==0.26.0
python-dotenv==1.0.0
openai==1.84.0